+++ src/Loader.php
@@ -9,7 +9,9 @@
 namespace FluentDOM\YAML\Symfony {
 
   use FluentDOM\Document;
+  use FluentDOM\DocumentFragment;
   use FluentDOM\Loadable;
+  use FluentDOM\Loader\Options;
   use FluentDOM\Loader\Supports;
 
   use Symfony\Component\Yaml\Parser;
@@ -33,26 +35,65 @@ namespace FluentDOM\YAML\Symfony {
      *
      * @param mixed $source
      * @param string $contentType
-     * @param array $options
+     * @param array|\Traversable|Options $options
      * @return Document|NULL
      */
-    public function load($source, $contentType, array $options = []) {
+    public function load($source, $contentType, $options = []) {
       if ($this->supports($contentType)) {
-        if (FALSE === strpos($source, "\n")) {
+        $settings = $this->getOptions($options);
+        $settings->isAllowed($sourceType = $settings->getSourceType($source));
+        switch ($sourceType) {
+        case Options::IS_FILE :
           $source = file_get_contents($source);
-        }
-        $parser = new Parser();
-        $yaml = $parser->parse($source);
-        if (!empty($yaml) || is_array($yaml)) {
-          $dom = new Document('1.0', 'UTF-8');
-          $dom->appendChild(
-            $root = $dom->createElementNS(self::XMLNS, 'json:json')
-          );
-          $this->transferTo($root, $yaml);
-          return $dom;
+        case Options::IS_STRING :
+        default :
+          $parser = new Parser();
+          $yaml = $parser->parse($source);
+          if (!empty($yaml) || is_array($yaml)) {
+            $document = new Document('1.0', 'UTF-8');
+            $document->appendChild(
+              $root = $document->createElementNS(self::XMLNS, 'json:json')
+            );
+            $this->transferTo($root, $yaml);
+            return $document;
+          }
         }
       }
       return NULL;
+    }
+
+    /**
+     * Load the YAML string into an DOMDocumentFragment
+     *
+     * @param mixed $source
+     * @param string $contentType
+     * @param array|\Traversable|Options $options
+     * @return DocumentFragment|NULL
+     */
+    public function loadFragment($source, $contentType, $options = []) {
+      if ($this->supports($contentType)) {
+        $parser = new Parser();
+        $yaml = $parser->parse($source);
+        if (!empty($yaml) || is_array($yaml)) {
+          $document = new Document('1.0', 'UTF-8');
+          $fragment = $document->createDocumentFragment();
+          $this->transferTo($fragment, $yaml);
+          return $fragment;
+        }
+      }
+      return NULL;
+    }
+
+    public function getOptions($options) {
+      $result = new Options(
+        $options,
+        [
+          Options::CB_IDENTIFY_STRING_SOURCE => function($source) {
+            return (FALSE !== strpos($source, "\n"));
+          }
+        ]
+      );
+      return $result;
     }
   }
 }